@@ -22,7 +22,7 @@ import (
 func (dr *DrainReconcile) handleNodeIdleNodeStateDrainingOrCompleted(ctx context.Context,
 	node *corev1.Node,
 	nodeNetworkState *sriovnetworkv1.SriovNetworkNodeState) (ctrl.Result, error) {
-	reqLogger := ctx.Value("logger").(logr.Logger).WithName("handleNodeIdleNodeStateDrainingOrCompleted")
+	reqLogger := ctx.Value(constants.LoggerContextKey).(logr.Logger).WithName("handleNodeIdleNodeStateDrainingOrCompleted")
 	completed, err := dr.drainer.CompleteDrainNode(ctx, node)
 	if err != nil {
 		reqLogger.Error(err, "failed to complete drain on node")
@@ -64,7 +64,7 @@ func (dr *DrainReconcile) handleNodeDrainOrReboot(ctx context.Context,
 	nodeNetworkState *sriovnetworkv1.SriovNetworkNodeState,
 	nodeDrainAnnotation,
 	nodeStateDrainAnnotationCurrent string) (ctrl.Result, error) {
-	reqLogger := ctx.Value("logger").(logr.Logger).WithName("handleNodeDrainOrReboot")
+	reqLogger := ctx.Value(constants.LoggerContextKey).(logr.Logger).WithName("handleNodeDrainOrReboot")
 	// nothing to do here we need to wait for the node to move back to idle
 	if nodeStateDrainAnnotationCurrent == constants.DrainComplete {
 		reqLogger.Info("node requested a drain and nodeState is on drain completed nothing todo")
@@ -138,7 +138,7 @@ func (dr *DrainReconcile) handleNodeDrainOrReboot(ctx context.Context,
 }
 
 func (dr *DrainReconcile) tryDrainNode(ctx context.Context, node *corev1.Node) (*reconcile.Result, error) {
-	reqLogger := ctx.Value("logger").(logr.Logger).WithName("tryDrainNode")
+	reqLogger := ctx.Value(constants.LoggerContextKey).(logr.Logger).WithName("tryDrainNode")
 
 	//critical section we need to check if we can start the draining
 	dr.drainCheckMutex.Lock()
@@ -208,7 +208,7 @@ func (dr *DrainReconcile) tryDrainNode(ctx context.Context, node *corev1.Node) (
 }
 
 func (dr *DrainReconcile) findNodePoolConfig(ctx context.Context, node *corev1.Node) (*sriovnetworkv1.SriovNetworkPoolConfig, []corev1.Node, error) {
-	logger := ctx.Value("logger").(logr.Logger).WithName("findNodePoolConfig")
+	logger := ctx.Value(constants.LoggerContextKey).(logr.Logger).WithName("findNodePoolConfig")
 	// get all the sriov network pool configs
 	npcl := &sriovnetworkv1.SriovNetworkPoolConfigList{}
 	err := dr.List(ctx, npcl)
