@@ -19,7 +19,6 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	uns "k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/wait"
@@ -323,7 +322,7 @@ var _ = Describe("SriovOperatorConfig controller", Ordered, func() {
 			})
 			Expect(err).NotTo(HaveOccurred())
 
-			config := &uns.Unstructured{}
+			config := &unstructured.Unstructured{}
 			config.SetGroupVersionKind(sriovnetworkv1.GroupVersion.WithKind("SriovOperatorConfig"))
 			config.SetName(consts.DefaultConfigName)
 			config.SetNamespace(testNamespace)
@@ -346,7 +345,7 @@ var _ = Describe("SriovOperatorConfig controller", Ordered, func() {
 			}, util.APITimeout, util.RetryInterval).Should(Succeed())
 
 			By("Verify default values have not been omitted")
-			obj := &uns.Unstructured{}
+			obj := &unstructured.Unstructured{}
 			obj.SetGroupVersionKind(sriovnetworkv1.GroupVersion.WithKind("SriovOperatorConfig"))
 			err = k8sClient.Get(context.Background(), types.NamespacedName{Namespace: testNamespace, Name: consts.DefaultConfigName}, obj)
 			Expect(err).NotTo(HaveOccurred())
@@ -449,7 +448,22 @@ var _ = Describe("SriovOperatorConfig controller", Ordered, func() {
 				return strings.Join(daemonSet.Spec.Template.Spec.Containers[0].Args, " ")
 			}, util.APITimeout*10, util.RetryInterval).Should(ContainSubstring("disable-plugins=mellanox"))
 		})
+		It("should render configDaemonEnvVars in sriov-network-config-daemon if provided in spec", func() {
+			config := &sriovnetworkv1.SriovOperatorConfig{}
+			Expect(k8sClient.Get(ctx, types.NamespacedName{Namespace: testNamespace, Name: "default"}, config)).NotTo(HaveOccurred())
 
+			config.Spec.ConfigDaemonEnvVars = map[string]string{"TEST_ENV_VAR": "test_value"}
+			err := k8sClient.Update(ctx, config)
+			Expect(err).NotTo(HaveOccurred())
+
+			Eventually(func(g Gomega) {
+				daemonSet := &appsv1.DaemonSet{}
+				err := k8sClient.Get(ctx, types.NamespacedName{Name: "sriov-network-config-daemon", Namespace: testNamespace}, daemonSet)
+				g.Expect(err).NotTo(HaveOccurred())
+				g.Expect(daemonSet.Spec.Template.Spec.Containers[0].Env).To(ContainElement(
+					corev1.EnvVar{Name: "TEST_ENV_VAR", Value: "test_value"}))
+			}, util.APITimeout*10, util.RetryInterval).Should(Succeed())
+		})
 		It("should render the resourceInjectorMatchCondition in the mutation if feature flag is enabled and block only pods with the networks annotation", func() {
 			By("set the feature flag")
 			config := &sriovnetworkv1.SriovOperatorConfig{}
