@@ -2,9 +2,22 @@ package featuregate
 
 import (
 	"fmt"
+	"maps"
 	"strings"
 	"sync"
+
+	"github.com/k8snetworkplumbingwg/sriov-network-operator/pkg/consts"
 )
+
+// DefaultFeatureStates contains the default states for the feature gates
+var DefaultFeatureStates = map[string]bool{
+	consts.ParallelNicConfigFeatureGate:                false,
+	consts.ResourceInjectorMatchConditionFeatureGate:   false,
+	consts.MetricsExporterFeatureGate:                  false,
+	consts.ManageSoftwareBridgesFeatureGate:            false,
+	consts.BlockDevicePluginUntilConfiguredFeatureGate: true,
+	consts.MellanoxFirmwareResetFeatureGate:            false,
+}
 
 // FeatureGate provides methods to check state of the feature
 type FeatureGate interface {
@@ -12,23 +25,34 @@ type FeatureGate interface {
 	// if feature name is unknown will always return false
 	IsEnabled(feature string) bool
 	// Init set state for the features from the provided map.
-	// completely removes the previous state
+	// The provided map is merged with the default features state.
 	Init(features map[string]bool)
 	// String returns string representation of the feature state
 	String() string
 }
 
-// New returns default implementation of the FeatureGate interface
+// New returns default implementation of the FeatureGate interface with the default features state
 func New() FeatureGate {
 	return &featureGate{
-		lock:  &sync.RWMutex{},
-		state: map[string]bool{},
+		lock:            &sync.RWMutex{},
+		state:           map[string]bool{},
+		defaultFeatures: DefaultFeatureStates,
+	}
+}
+
+// NewWithDefaultFeatures returns a new FeatureGate with the default features state explicitly set
+func NewWithDefaultFeatures(defaultFeatures map[string]bool) FeatureGate {
+	return &featureGate{
+		lock:            &sync.RWMutex{},
+		state:           map[string]bool{},
+		defaultFeatures: defaultFeatures,
 	}
 }
 
 type featureGate struct {
-	lock  *sync.RWMutex
-	state map[string]bool
+	lock            *sync.RWMutex
+	state           map[string]bool
+	defaultFeatures map[string]bool
 }
 
 // IsEnabled returns state of the feature,
@@ -40,17 +64,19 @@ func (fg *featureGate) IsEnabled(feature string) bool {
 }
 
 // Init set state for the features from the provided map.
-// completely removes the previous state
+// The provided features override the default values.
 func (fg *featureGate) Init(features map[string]bool) {
 	fg.lock.Lock()
 	defer fg.lock.Unlock()
-	fg.state = make(map[string]bool, len(features))
-	for k, v := range features {
-		fg.state[k] = v
+	state := maps.Clone(fg.defaultFeatures)
+	if state == nil {
+		state = map[string]bool{}
 	}
+	maps.Copy(state, features)
+	fg.state = state
 }
 
-// String returns string representation of the feature state
+// String returns string representation of the features state
 func (fg *featureGate) String() string {
 	fg.lock.RLock()
 	defer fg.lock.RUnlock()
