@@ -10,13 +10,13 @@ import (
 var _ = Describe("GUID", func() {
 	It("should parse and process GUIDs correctly", func() {
 		guidStr := "00:01:02:03:04:05:06:08"
-		nextGuidStr := "00:01:02:03:04:05:06:09"
+		nextGUIDStr := "00:01:02:03:04:05:06:09"
 
 		guid, err := ParseGUID(guidStr)
 		Expect(err).NotTo(HaveOccurred())
 
 		Expect(guid.String()).To(Equal(guidStr))
-		Expect((guid + 1).String()).To(Equal(nextGuidStr))
+		Expect((guid + 1).String()).To(Equal(nextGUIDStr))
 	})
 	It("should represent GUID as HW address", func() {
 		guidStr := "00:01:02:03:04:05:06:08"
