@@ -241,11 +241,11 @@ func (n *network) GetNetDevNodeGUID(pciAddr string) string {
 }
 
 func (n *network) GetNetDevLinkSpeed(ifaceName string) string {
-	log.Log.V(2).Info("GetNetDevLinkSpeed(): get LinkSpeed", "device", ifaceName)
+	funcLog := log.Log.WithValues("device", ifaceName)
 	speedFilePath := filepath.Join(vars.FilesystemRoot, consts.SysClassNet, ifaceName, "speed")
 	data, err := os.ReadFile(speedFilePath)
 	if err != nil {
-		log.Log.Error(err, "GetNetDevLinkSpeed(): fail to read Link Speed file", "path", speedFilePath)
+		funcLog.Info("GetNetDevLinkSpeed(): WARNING: fail to read Link Speed file", "path", speedFilePath, "error", err)
 		return ""
 	}
 
@@ -458,7 +458,7 @@ func (n *network) SetRDMASubsystem(mode string) error {
 	}
 	config := fmt.Sprintf("# This file is managed by sriov-network-operator do not edit.\noptions ib_core netns_mode=%d\n", modeValue)
 
-	err := os.WriteFile(path, []byte(config), 0644)
+	err := os.WriteFile(path, []byte(config), 0o644)
 	if err != nil {
 		log.Log.Error(err, "SetRDMASubsystem(): failed to write sriov_network_operator_modules_config.conf")
 		return fmt.Errorf("failed to write sriov_network_operator_modules_config.conf: %v", err)
