@@ -1,6 +1,7 @@
 package sriov
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"os"
@@ -28,8 +29,8 @@ import (
 )
 
 type interfaceToConfigure struct {
-	iface       sriovnetworkv1.Interface
-	ifaceStatus sriovnetworkv1.InterfaceExt
+	Iface       sriovnetworkv1.Interface
+	IfaceStatus sriovnetworkv1.InterfaceExt
 }
 
 type sriov struct {
@@ -173,7 +174,7 @@ func (s *sriov) VFIsReady(pciAddr string) (netlink.Link, error) {
 	log.Log.Info("VFIsReady()", "device", pciAddr)
 	var err error
 	var vfLink netlink.Link
-	err = wait.PollImmediate(time.Second, 10*time.Second, func() (bool, error) {
+	err = wait.PollUntilContextTimeout(context.Background(), time.Second, 10*time.Second, true, func(ctx context.Context) (bool, error) {
 		vfIndex, err := s.networkHelper.GetInterfaceIndex(pciAddr)
 		if err != nil {
 			log.Log.Error(err, "VFIsReady(): invalid index number", "device", pciAddr)
@@ -208,6 +209,15 @@ func (s *sriov) SetVfAdminMac(vfAddr string, pfLink, vfLink netlink.Link) error 
 	return nil
 }
 
+// getDeviceClass parses the device class from the device
+func getDeviceClass(device *ghw.PCIDevice) (int64, error) {
+	devClass, err := strconv.ParseInt(device.Class.ID, 16, 64)
+	if err != nil {
+		return 0, fmt.Errorf("unable to parse device class: %v", err)
+	}
+	return devClass, nil
+}
+
 func (s *sriov) DiscoverSriovDevices(storeManager store.ManagerInterface) ([]sriovnetworkv1.InterfaceExt, error) {
 	log.Log.V(2).Info("DiscoverSriovDevices")
 	pfList := []sriovnetworkv1.InterfaceExt{}
@@ -223,7 +233,7 @@ func (s *sriov) DiscoverSriovDevices(storeManager store.ManagerInterface) ([]sri
 	}
 
 	for _, device := range devices {
-		devClass, err := strconv.ParseInt(device.Class.ID, 16, 64)
+		devClass, err := getDeviceClass(device)
 		if err != nil {
 			log.Log.Error(err, "DiscoverSriovDevices(): unable to parse device class, skipping",
 				"device", device)
@@ -305,6 +315,63 @@ func (s *sriov) DiscoverSriovDevices(storeManager store.ManagerInterface) ([]sri
 				}
 			}
 		}
+		pfList = append(pfList, iface)
+	}
+
+	return pfList, nil
+}
+
+func (s *sriov) DiscoverSriovVirtualDevices() ([]sriovnetworkv1.InterfaceExt, error) {
+	log.Log.V(2).Info("DiscoverSriovVirtualDevices()")
+	pfList := []sriovnetworkv1.InterfaceExt{}
+
+	pci, err := s.ghwLib.PCI()
+	if err != nil {
+		return nil, fmt.Errorf("DiscoverSriovVirtualDevices(): error getting PCI info: %v", err)
+	}
+
+	devices := pci.Devices
+	if len(devices) == 0 {
+		return nil, fmt.Errorf("DiscoverSriovVirtualDevices(): could not retrieve PCI devices")
+	}
+
+	for _, device := range devices {
+		devClass, err := getDeviceClass(device)
+		if err != nil {
+			log.Log.Error(err, "DiscoverSriovVirtualDevices(): unable to parse device class for device, skipping",
+				"device", device)
+			continue
+		}
+		if devClass != consts.NetClass {
+			// Not network device
+			continue
+		}
+
+		driver, err := s.dputilsLib.GetDriverName(device.Address)
+		if err != nil {
+			log.Log.Error(err, "DiscoverSriovVirtualDevices(): unable to parse device driver for device, skipping",
+				"device", device)
+			continue
+		}
+		iface := sriovnetworkv1.InterfaceExt{
+			PciAddress: device.Address,
+			Driver:     driver,
+			Vendor:     device.Vendor.ID,
+			DeviceID:   device.Product.ID,
+		}
+
+		if mtu := s.networkHelper.GetNetdevMTU(device.Address); mtu > 0 {
+			iface.Mtu = mtu
+		}
+		if name := s.networkHelper.TryToGetVirtualInterfaceName(device.Address); name != "" {
+			iface.Name = name
+			if macAddr := s.networkHelper.GetNetDevMac(name); macAddr != "" {
+				iface.Mac = macAddr
+			}
+			iface.LinkSpeed = s.networkHelper.GetNetDevLinkSpeed(name)
+			iface.LinkType = s.GetLinkType(name)
+		}
+
 		pfList = append(pfList, iface)
 	}
 
@@ -412,7 +479,7 @@ func (s *sriov) checkExternallyManagedPF(iface *sriovnetworkv1.Interface) error 
 			"functions %d but the policy is configured as ExternallyManaged for device %s",
 			iface.NumVfs, currentNumVfs, iface.PciAddress)
 		log.Log.Error(nil, errMsg)
-		return fmt.Errorf(errMsg)
+		return errors.New(errMsg)
 	}
 	currentEswitchMode := s.GetNicSriovMode(iface.PciAddress)
 	expectedEswitchMode := sriovnetworkv1.GetEswitchModeFromSpec(iface)
@@ -420,7 +487,7 @@ func (s *sriov) checkExternallyManagedPF(iface *sriovnetworkv1.Interface) error 
 		errMsg := fmt.Sprintf("checkExternallyManagedPF(): requested ESwitchMode mode \"%s\" is not equal to configured \"%s\" "+
 			"but the policy is configured as ExternallyManaged for device %s", expectedEswitchMode, currentEswitchMode, iface.PciAddress)
 		log.Log.Error(nil, errMsg)
-		return fmt.Errorf(errMsg)
+		return errors.New(errMsg)
 	}
 	currentMtu := s.networkHelper.GetNetdevMTU(iface.PciAddress)
 	if iface.Mtu > 0 && iface.Mtu > currentMtu {
@@ -658,7 +725,7 @@ func (s *sriov) getConfigureAndReset(storeManager store.ManagerInterface, interf
 				}
 				iface := iface
 				ifaceStatus := ifaceStatus
-				toBeConfigured = append(toBeConfigured, interfaceToConfigure{iface: iface, ifaceStatus: ifaceStatus})
+				toBeConfigured = append(toBeConfigured, interfaceToConfigure{Iface: iface, IfaceStatus: ifaceStatus})
 			}
 		}
 
@@ -679,12 +746,12 @@ func (s *sriov) configSriovInterfacesInParallel(storeManager store.ManagerInterf
 		interfacesToConfigure += 1
 		go func(iface *interfaceToConfigure) {
 			var err error
-			if err = s.configSriovDevice(&iface.iface, skipVFConfiguration); err != nil {
-				log.Log.Error(err, "configSriovInterfacesInParallel(): fail to configure sriov interface. resetting interface.", "address", iface.iface.PciAddress)
-				if iface.iface.ExternallyManaged {
+			if err = s.configSriovDevice(&iface.Iface, skipVFConfiguration); err != nil {
+				log.Log.Error(err, "configSriovInterfacesInParallel(): fail to configure sriov interface. resetting interface.", "address", iface.Iface.PciAddress)
+				if iface.Iface.ExternallyManaged {
 					log.Log.V(2).Info("configSriovInterfacesInParallel(): skipping device reset as the nic is marked as externally created")
 				} else {
-					if resetErr := s.ResetSriovDevice(iface.ifaceStatus); resetErr != nil {
+					if resetErr := s.ResetSriovDevice(iface.IfaceStatus); resetErr != nil {
 						log.Log.Error(resetErr, "configSriovInterfacesInParallel(): failed to reset on error SR-IOV interface")
 						err = resetErr
 					}
@@ -693,7 +760,7 @@ func (s *sriov) configSriovInterfacesInParallel(storeManager store.ManagerInterf
 			errChannel <- err
 		}(&interfaces[ifaceIndex])
 		// Save the PF status to the host
-		err := storeManager.SaveLastPfAppliedStatus(&iface.iface)
+		err := storeManager.SaveLastPfAppliedStatus(&iface.Iface)
 		if err != nil {
 			log.Log.Error(err, "configSriovInterfacesInParallel(): failed to save PF applied config to host")
 			return err
@@ -743,12 +810,12 @@ func (s *sriov) resetSriovInterfacesInParallel(storeManager store.ManagerInterfa
 func (s *sriov) configSriovInterfaces(storeManager store.ManagerInterface, interfaces []interfaceToConfigure, skipVFConfiguration bool) error {
 	log.Log.V(2).Info("configSriovInterfaces(): start sriov configuration")
 	for _, iface := range interfaces {
-		if err := s.configSriovDevice(&iface.iface, skipVFConfiguration); err != nil {
-			log.Log.Error(err, "configSriovInterfaces(): fail to configure sriov interface. resetting interface.", "address", iface.iface.PciAddress)
-			if iface.iface.ExternallyManaged {
+		if err := s.configSriovDevice(&iface.Iface, skipVFConfiguration); err != nil {
+			log.Log.Error(err, "configSriovInterfaces(): fail to configure sriov interface. resetting interface.", "address", iface.Iface.PciAddress)
+			if iface.Iface.ExternallyManaged {
 				log.Log.V(2).Info("configSriovInterfaces(): skipping device reset as the nic is marked as externally created")
 			} else {
-				if resetErr := s.ResetSriovDevice(iface.ifaceStatus); resetErr != nil {
+				if resetErr := s.ResetSriovDevice(iface.IfaceStatus); resetErr != nil {
 					log.Log.Error(resetErr, "configSriovInterfaces(): failed to reset on error SR-IOV interface")
 				}
 			}
@@ -756,7 +823,7 @@ func (s *sriov) configSriovInterfaces(storeManager store.ManagerInterface, inter
 		}
 
 		// Save the PF status to the host
-		err := storeManager.SaveLastPfAppliedStatus(&iface.iface)
+		err := storeManager.SaveLastPfAppliedStatus(&iface.Iface)
 		if err != nil {
 			log.Log.Error(err, "configSriovInterfaces(): failed to save PF applied config to host")
 			return err
@@ -843,49 +910,103 @@ func (s *sriov) checkForConfigAndReset(ifaceStatus sriovnetworkv1.InterfaceExt, 
 	return nil
 }
 
-func (s *sriov) ConfigSriovDeviceVirtual(iface *sriovnetworkv1.Interface) error {
-	log.Log.V(2).Info("ConfigSriovDeviceVirtual(): config interface", "address", iface.PciAddress, "config", iface)
-	// Config VFs
-	if iface.NumVfs > 0 {
-		if iface.NumVfs > 1 {
-			log.Log.Error(nil, "ConfigSriovDeviceVirtual(): in a virtual environment, only one VF per interface",
-				"numVfs", iface.NumVfs)
-			return errors.New("NumVfs > 1")
-		}
-		if len(iface.VfGroups) != 1 {
-			log.Log.Error(nil, "ConfigSriovDeviceVirtual(): missing VFGroup")
-			return errors.New("NumVfs != 1")
-		}
-		addr := iface.PciAddress
-		log.Log.V(2).Info("ConfigSriovDeviceVirtual()", "address", addr)
-		driver := ""
-		vfID := 0
-		for _, group := range iface.VfGroups {
-			log.Log.V(2).Info("ConfigSriovDeviceVirtual()", "group", group)
-			if sriovnetworkv1.IndexInRange(vfID, group.VfRange) {
-				log.Log.V(2).Info("ConfigSriovDeviceVirtual()", "indexInRange", vfID)
-				if sriovnetworkv1.StringInArray(group.DeviceType, vars.DpdkDrivers) {
-					log.Log.V(2).Info("ConfigSriovDeviceVirtual()", "driver", group.DeviceType)
-					driver = group.DeviceType
+func (s *sriov) ConfigSriovDevicesVirtual(storeManager store.ManagerInterface, interfaces []sriovnetworkv1.Interface, ifaceStatuses []sriovnetworkv1.InterfaceExt) error {
+	toBeConfigured, toBeResetted, err := s.getConfigureAndReset(storeManager, interfaces, ifaceStatuses)
+	if err != nil {
+		log.Log.Error(err, "ConfigSriovDevicesVirtual(): cannot get a list of interfaces to configure")
+		return fmt.Errorf("cannot get a list of interfaces to configure")
+	}
+	log.Log.V(2).Info("ConfigSriovDevicesVirtual(): configuration to be done", "toBeConfigured", toBeConfigured, "toBeResetted", toBeResetted)
+
+	for _, ifaceToConfigure := range toBeConfigured {
+		if ifaceToConfigure.Iface.NumVfs > 0 {
+			if ifaceToConfigure.Iface.NumVfs > 1 {
+				log.Log.Error(nil, "ConfigSriovDeviceVirtual(): in a virtual environment, only one VF per interface",
+					"numVfs", ifaceToConfigure.Iface.NumVfs)
+				return errors.New("NumVfs > 1")
+			}
+			// Validate the VFGroups we need to have only one VFGroup
+			if len(ifaceToConfigure.Iface.VfGroups) != 1 {
+				log.Log.Error(nil, "ConfigSriovDeviceVirtual(): unexpected number of VFGroups", "vfGroups", ifaceToConfigure.Iface.VfGroups)
+				return errors.New("unexpected number of VFGroups")
+			}
+
+			addr := ifaceToConfigure.Iface.PciAddress
+			log.Log.V(2).Info("ConfigSriovDeviceVirtual()", "pciAddress", addr)
+			driver := ""
+			vfID := 0
+
+			group := ifaceToConfigure.Iface.VfGroups[0]
+			// Validate the VF ID is in the range for the VFGroup
+			if !sriovnetworkv1.IndexInRange(vfID, group.VfRange) {
+				log.Log.Error(nil, "ConfigSriovDeviceVirtual(): VF ID is not in the range", "vfID", vfID, "vfRange", group.VfRange)
+				return errors.New("VF ID is not in the range")
+			}
+
+			// check if we need to bind to a DPDK driver
+			if sriovnetworkv1.StringInArray(group.DeviceType, vars.DpdkDrivers) {
+				log.Log.V(2).Info("ConfigSriovDeviceVirtual()", "driver", group.DeviceType)
+				driver = group.DeviceType
+			}
+
+			if driver == "" {
+				log.Log.V(2).Info("ConfigSriovDeviceVirtual(): bind default")
+				if err := s.kernelHelper.BindDefaultDriver(addr); err != nil {
+					log.Log.Error(err, "ConfigSriovDeviceVirtual(): fail to bind default driver", "device", addr)
+					return err
 				}
-				break
+			} else {
+				log.Log.V(2).Info("ConfigSriovDeviceVirtual(): bind driver", "driver", driver)
+				if err := s.kernelHelper.BindDpdkDriver(addr, driver); err != nil {
+					log.Log.Error(err, "ConfigSriovDeviceVirtual(): fail to bind driver for device",
+						"driver", driver, "device", addr)
+					return err
+				}
 			}
-		}
-		if driver == "" {
-			log.Log.V(2).Info("ConfigSriovDeviceVirtual(): bind default")
-			if err := s.kernelHelper.BindDefaultDriver(addr); err != nil {
-				log.Log.Error(err, "ConfigSriovDeviceVirtual(): fail to bind default driver", "device", addr)
-				return err
-			}
-		} else {
-			log.Log.V(2).Info("ConfigSriovDeviceVirtual(): bind driver", "driver", driver)
-			if err := s.kernelHelper.BindDpdkDriver(addr, driver); err != nil {
-				log.Log.Error(err, "ConfigSriovDeviceVirtual(): fail to bind driver for device",
-					"driver", driver, "device", addr)
+			// Save the PF status to the host
+			err := storeManager.SaveLastPfAppliedStatus(&ifaceToConfigure.Iface)
+			if err != nil {
+				log.Log.Error(err, "configSriovInterfaces(): failed to save PF applied config to host")
 				return err
 			}
 		}
 	}
+
+	for _, iface := range toBeResetted {
+		// check if the vf were created by the sriov operator
+		pfStatus, exist, err := storeManager.LoadPfsStatus(iface.PciAddress)
+		if err != nil {
+			log.Log.Error(err, "ConfigSriovDeviceVirtual(): failed to load info about PF status for device",
+				"address", iface.PciAddress)
+			return err
+		}
+		if !exist {
+			log.Log.V(2).Info("ConfigSriovDeviceVirtual(): PF name with pci address has VFs configured but they weren't created by the sriov operator. Skipping the device reset",
+				"pf-name", iface.Name,
+				"address", iface.PciAddress)
+			continue
+		}
+		if pfStatus.ExternallyManaged {
+			log.Log.V(2).Info("ConfigSriovDeviceVirtual(): PF name with pci address was externally created skipping the device reset",
+				"pf-name", iface.Name,
+				"address", iface.PciAddress)
+			continue
+		}
+
+		log.Log.V(2).Info("ConfigSriovDeviceVirtual(): bind default")
+		if err := s.kernelHelper.BindDefaultDriver(iface.PciAddress); err != nil {
+			log.Log.Error(err, "ConfigSriovDeviceVirtual(): fail to bind default driver", "device", iface.PciAddress)
+			return err
+		}
+
+		// Remove the PF status from the host
+		err = storeManager.RemovePfAppliedStatus(iface.PciAddress)
+		if err != nil {
+			log.Log.Error(err, "configSriovInterfaces(): failed to remove PF applied config from host")
+			return err
+		}
+	}
+
 	return nil
 }
 
@@ -1033,6 +1154,71 @@ func (s *sriov) setEswitchModeAndNumVFs(pciAddr string, desiredEswitchMode strin
 	return fn(pciAddr, desiredEswitchMode, numVFs)
 }
 
+func (s *sriov) waitForVFLinks(pciAddr string, expectedNum int, maxTimeout time.Duration) error {
+	// VF destruction is asynchronous and safe to ignore
+	if expectedNum == 0 {
+		log.Log.V(2).Info("waitForVFLinks(): skipping wait for cleanup (expected 0 VFs) - async destroy", "device", pciAddr)
+		return nil
+	}
+
+	log.Log.V(2).Info("waitForVFLinks(): waiting for VF symlinks",
+		"device", pciAddr,
+		"expected", expectedNum,
+		"maxTimeout", maxTimeout)
+
+	err := wait.PollUntilContextTimeout(
+		context.Background(),
+		time.Second, // poll interval
+		maxTimeout,  // overall timeout
+		true,        // run immediately
+		func(ctx context.Context) (bool, error) {
+			vfAddrs, err := s.dputilsLib.GetVFList(pciAddr)
+			if err != nil {
+				log.Log.V(2).Info("waitForVFLinks(): GetVFList failed, retrying", "err", err)
+				return false, nil
+			}
+
+			if len(vfAddrs) < expectedNum {
+				return false, nil
+			}
+
+			// Check all have physfn symlink
+			for _, vfAddr := range vfAddrs {
+				linkPath := filepath.Join(
+					vars.FilesystemRoot,
+					consts.SysBusPciDevices,
+					vfAddr,
+					"physfn",
+				)
+
+				if _, err := os.Lstat(linkPath); err != nil {
+					log.Log.V(2).Info(
+						"waitForVFLinks(): physfn symlink missing",
+						"vf", vfAddr,
+						"err", err,
+					)
+					return false, nil
+				}
+			}
+
+			log.Log.V(2).Info("waitForVFLinks(): all expected VF symlinks ready")
+			return true, nil
+		},
+	)
+
+	if err != nil {
+		return fmt.Errorf(
+			"timeout waiting for %d VF symlinks on %s (max %v): %w",
+			expectedNum,
+			pciAddr,
+			maxTimeout,
+			err,
+		)
+	}
+
+	return nil
+}
+
 // setEswitchModeAndNumVFsMlx configures PF eSwitch and sriov_numvfs in the following order:
 // a. set eSwitchMode to legacy
 // b. set the desired number of Virtual Functions
@@ -1059,6 +1245,9 @@ func (s *sriov) setEswitchModeAndNumVFsMlx(pciAddr string, desiredEswitchMode st
 		}
 	}
 	if err := s.SetSriovNumVfs(pciAddr, numVFs); err != nil {
+		return err
+	}
+	if err := s.waitForVFLinks(pciAddr, numVFs, 120*time.Second); err != nil {
 		return err
 	}
 
