@@ -11,7 +11,6 @@ import (
 	. "github.com/onsi/gomega"
 
 	"github.com/k8snetworkplumbingwg/sriov-network-operator/pkg/consts"
-	constants "github.com/k8snetworkplumbingwg/sriov-network-operator/pkg/consts"
 	netlinkMock "github.com/k8snetworkplumbingwg/sriov-network-operator/pkg/host/internal/lib/netlink/mock"
 	hostMock "github.com/k8snetworkplumbingwg/sriov-network-operator/pkg/host/mock"
 	"github.com/k8snetworkplumbingwg/sriov-network-operator/pkg/host/types"
@@ -37,7 +36,7 @@ var _ = Describe("VDPA", func() {
 	})
 	Context("CreateVDPADevice", func() {
 		callFunc := func() error {
-			return v.CreateVDPADevice("0000:d8:00.2", constants.VdpaTypeVhost)
+			return v.CreateVDPADevice("0000:d8:00.2", consts.VdpaTypeVhost)
 		}
 		It("Created", func() {
 			libMock.EXPECT().VDPAGetDevByName("vdpa:0000:d8:00.2").Return(nil, syscall.ENODEV)
