@@ -136,12 +136,16 @@ type SriovInterface interface {
 	ResetSriovDevice(ifaceStatus sriovnetworkv1.InterfaceExt) error
 	// DiscoverSriovDevices returns a list of all the available SR-IOV capable network interfaces on the system
 	DiscoverSriovDevices(storeManager store.ManagerInterface) ([]sriovnetworkv1.InterfaceExt, error)
+	// DiscoverSriovVirtualDevices returns a list of all the available SR-IOV VF network interfaces on the system.
+	// Only supported on virtual environments where we don't have the physical function.
+	DiscoverSriovVirtualDevices() ([]sriovnetworkv1.InterfaceExt, error)
 	// ConfigSriovInterfaces configure multiple SR-IOV devices with the desired configuration
 	// if skipVFConfiguration flag is set, the function will configure PF and create VFs on it, but will skip VFs configuration
 	ConfigSriovInterfaces(storeManager store.ManagerInterface, interfaces []sriovnetworkv1.Interface,
 		ifaceStatuses []sriovnetworkv1.InterfaceExt, skipVFConfiguration bool) error
-	// ConfigSriovInterfaces configure virtual functions for virtual environments with the desired configuration
-	ConfigSriovDeviceVirtual(iface *sriovnetworkv1.Interface) error
+	// ConfigSriovDevicesVirtual configure virtual functions for virtual environments with the desired configuration
+	ConfigSriovDevicesVirtual(storeManager store.ManagerInterface, interfaces []sriovnetworkv1.Interface,
+		ifaceStatuses []sriovnetworkv1.InterfaceExt) error
 }
 
 type UdevInterface interface {
