@@ -33,8 +33,8 @@ var (
 
 func WaitForSriovNetworkNodeStateReady(nodeState *sriovnetworkv1.SriovNetworkNodeState, client client.Client, namespace, name string, retryInterval, timeout time.Duration) error {
 	time.Sleep(30 * time.Second)
-	err := wait.PollImmediate(retryInterval, timeout, func() (done bool, err error) {
-		ctx, cancel := goctx.WithTimeout(goctx.Background(), APITimeout)
+	err := wait.PollUntilContextTimeout(goctx.Background(), retryInterval, timeout, true, func(pollCtx goctx.Context) (done bool, err error) {
+		ctx, cancel := goctx.WithTimeout(pollCtx, APITimeout)
 		defer cancel()
 		err = client.Get(ctx, types.NamespacedName{Name: name, Namespace: namespace}, nodeState)
 		if err != nil {
@@ -57,8 +57,8 @@ func WaitForSriovNetworkNodeStateReady(nodeState *sriovnetworkv1.SriovNetworkNod
 }
 
 func WaitForDaemonSetReady(ds *appsv1.DaemonSet, client client.Client, namespace, name string, retryInterval, timeout time.Duration) error {
-	err := wait.PollImmediate(retryInterval, timeout, func() (done bool, err error) {
-		ctx, cancel := goctx.WithTimeout(goctx.Background(), APITimeout)
+	err := wait.PollUntilContextTimeout(goctx.Background(), retryInterval, timeout, true, func(pollCtx goctx.Context) (done bool, err error) {
+		ctx, cancel := goctx.WithTimeout(pollCtx, APITimeout)
 		defer cancel()
 		err = client.Get(ctx, types.NamespacedName{Name: name, Namespace: namespace}, ds)
 		if err != nil {
@@ -82,8 +82,8 @@ func WaitForDaemonSetReady(ds *appsv1.DaemonSet, client client.Client, namespace
 }
 
 func WaitForNamespacedObject(obj client.Object, client client.Client, namespace, name string, retryInterval, timeout time.Duration) error {
-	err := wait.PollImmediate(retryInterval, timeout, func() (done bool, err error) {
-		ctx, cancel := goctx.WithTimeout(goctx.Background(), APITimeout)
+	err := wait.PollUntilContextTimeout(goctx.Background(), retryInterval, timeout, true, func(pollCtx goctx.Context) (done bool, err error) {
+		ctx, cancel := goctx.WithTimeout(pollCtx, APITimeout)
 		defer cancel()
 		err = client.Get(ctx, types.NamespacedName{Name: name, Namespace: namespace}, obj)
 		if err != nil {
@@ -103,8 +103,8 @@ func WaitForNamespacedObject(obj client.Object, client client.Client, namespace,
 }
 
 func WaitForNamespacedObjectDeleted(obj client.Object, client client.Client, namespace, name string, retryInterval, timeout time.Duration) error {
-	err := wait.PollImmediate(retryInterval, timeout, func() (done bool, err error) {
-		ctx, cancel := goctx.WithTimeout(goctx.Background(), APITimeout)
+	err := wait.PollUntilContextTimeout(goctx.Background(), retryInterval, timeout, true, func(pollCtx goctx.Context) (done bool, err error) {
+		ctx, cancel := goctx.WithTimeout(pollCtx, APITimeout)
 		defer cancel()
 		err = client.Get(ctx, types.NamespacedName{Name: name, Namespace: namespace}, obj)
 		if err != nil {
